requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.8
django-dirtyfields==1.9.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.8.0